erro filho retornando um tipo primitivo maior q do pai

+++ OCPLearning/src/pkgs/pkg1/Animal.java
@@ -5,5 +5,5 @@ import pkgs.pkg2.Omega;
 
 public class Animal{
 	void locomover(){}
-	int alimentar(){return 0;}
+	byte alimentar(){return 0;}
 }
